requirements.txt
absl-py==2.0.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.5
decorator==5.1.1
Django==4.2.4
efficientnet-pytorch==0.7.1
executing==2.0.1
extra-streamlit-components==0.1.71
filelock==3.12.4
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2023.9.2
gast==0.5.4
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
googletrans==3.1.0a0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.14.0
kiwisolver==1.4.5
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-text-splitters==0.0.1
langsmith==0.1.52
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
orjson==3.10.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.3.2
scikit-plot==0.3.7
scipy==1.11.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.32.2
streamlit-authenticator==0.3.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.1
torch==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.4
youtube-transcript-api==0.6.2
